Consolidate conflict tests into a single spec file
